requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0